script/requirements.txt
numpy
matplotlib
pandas
pyyaml
notebook
seaborn
psptool
